refactor: unify DetectedLabels into single canonical type in pkg/shared/types
Consolidate three duplicate DetectedLabels struct definitions
(enrichment, models, shared/types) into pkg/shared/types as the
single source of truth. Both enrichment and models now use type
aliases, eliminating implicit JSON contract coupling identified
in the PR review (Architecture MEDIUM finding).

Changes:
- Add omitempty to bool fields on the canonical type (false omitted)
- Move Scan/Value/IsEmpty/NewDetectedLabels methods to shared type
- Replace enrichment.DetectedLabels struct with type alias
- Replace models.DetectedLabels struct with type alias
- Remove dead validate struct tags (never called at runtime)
