+++ pkg/kubernautagent/types/types.go
@@ -168,7 +168,7 @@ type SignalContext struct {
 	SignalAnnotations map[string]string `json:"signal_annotations,omitempty"`
 	SignalLabels      map[string]string `json:"signal_labels,omitempty"`
 
-	// DetectedLabelsJSON is a pre-marshaled JSON string of enrichment.DetectedLabels,
+	// DetectedLabelsJSON is a pre-marshaled JSON string of sharedtypes.DetectedLabels,
 	// forwarded to DS catalog queries to activate GitOps-aware scoring. Issue #1052 / BR-AI-056.
 	DetectedLabelsJSON string `json:"detected_labels_json,omitempty"`
